优化已各bug,max与mins校验long id时，异常问题

+++ src/main/java/com/github/ddm4j/api/document/bean/HeadVo.java
@@ -5,8 +5,8 @@ import java.io.Serializable;
 public class HeadVo extends ParamBaseVo implements Serializable {
 	private static final long serialVersionUID = 1L;
 	private String regexp;
-	private Integer min;
-	private Integer max;
+	private Long min;
+	private Long max;
 
 	public String getRegexp() {
 		return regexp;
@@ -16,19 +16,19 @@ public class HeadVo extends ParamBaseVo implements Serializable {
 		this.regexp = regexp;
 	}
 
-	public Integer getMin() {
+	public Long getMin() {
 		return min;
 	}
 
-	public void setMin(Integer min) {
+	public void setMin(Long min) {
 		this.min = min;
 	}
 
-	public Integer getMax() {
+	public Long getMax() {
 		return max;
 	}
 
-	public void setMax(Integer max) {
+	public void setMax(Long max) {
 		this.max = max;
 	}
